install EknServicesMultiplexer on all images
The new EknServicesMultiplexer flatpak eliminates the need to examine the
installed SDK versions to decide which versions of EknServices should be
installed. Instead, we install this multiplexer app and ensure it is up
to date, and this exports the services for all supported SDK versions which
are installed.

Use apps_add_mandatory as the variable so that if any personalities or
other config files do apps_del_foo = ${apps_del_default} they don't remove
the Flatpak which is marked as auto-install and would fail in the
60-flatpak-autoinstall-counters.chroot.

https://phabricator.endlessm.com/T22062