config: Disambiguate ostree ref and branch
In our usage here and in the ostree builder, branch refers to the OS
branch. Furthermore, although ostree often uses branch for the HEAD
commit pointers, they are more often is referred to as refs. Rework the
ostree config section to match that usage so that when we're working
with refs in the code, they're referred to as refs. This will also make
it more consistent when overrides for the refs are needed.

https://phabricator.endlessm.com/T22056