image: Introduce 'cros' partition table flag
Currently we have two ways to boot a rk3288 image on the Jerry Veyron
chromebook: (1) chainloading U-Boot (on the SD or eMMC) from depthcharge
or (2) boot Endless OS from U-Boot stored into the SPI.

Method (1) requires that the disk image layout is prepared in a special
way, reserving a partition for U-Boot and using 'cgpt' to create GPT
partitions with Chromium OS extensions.

Currently eos-image-builder only supports (1) and we basically need to
add support also for (2).

We do that by introducing the EIB_IMAGE_RK3288_CROS flag. When this flag
is defined at image generation time, we enforce (1). When the flag is
missing we generate the image with the regular Endless GPT and create
also the 'boot-spi.zip' file containing the SPI image to flash.

The image that is going to be flashed in the factory is the one with a
regular GPT partition table and U-Boot flashed into the SPI. The CROS
type image is only useful during development when we want to quickly
switch from CrOS to Endless OS at boot for bebugging / testing reasons.
We can drop support for it once the image is ready for production.

Signed-off-by: Carlo Caione <carlo@endlessm.com>