image: Call flatpak with --system
Recent flatpak changed behavior when called as root. I believe it tries
to setup a user installation at /root/.local/share/flatpak instead of
operating exclusively on the system installation. This is a problem in
the chroot because /root is a symlink to a nonexistent /var/roothome. It
ends up in a strange assert in libglnx:

  + flatpak remote-modify --disable eos
  **
  flatpak:ERROR:libglnx/glnx-shutil.c:147:mkdir_p_at_internal: assertion failed: (!did_recurse)

We don't want a user installation in /root anyways, so specify --system
to flatpak so it only operates on the system repo.

https://phabricator.endlessm.com/T22345