FlatpakRemote: Disable strict metadata parsing
Disable strict parsing so that it doesn't error on duplicate sections or
options, which flatpak-builder apparently generates sometimes. This has
recently happened with flathub's runtime/org.gnome.Platform/arm/3.24,
which has the following at the bottom of the metadata:

  [Build]
  built-extensions=org.gnome.Sdk.Locale;org.gnome.Sdk.Debug;org.gnome.Sdk.Docs;

  [Extension org.gnome.Platform.Locale]
  directory=share/runtime/locale
  autodelete=true
  locale-subset=true

  [Build]
  built-extensions=org.gnome.Platform.Locale;

This causes the following error in ConfigParser's default strict mode:

  configparser.DuplicateSectionError: While reading from '<string>'
  [line 64]: section 'Build' already exists

While this is most likely a flatpak-builder bug when it constructs the
metadata and should be fixed and deployed to flathub's builders, we
should try to carry on in the image builder.

https://phabricator.endlessm.com/T22435