config: Specify stable as default eos-sdk branch
When eos-sdk only contained runtimes, it didn't make sense to specify a
default branch since a runtime branch is always supplied. However, now
eos-sdk has some apps that are built on both master and stable.

Set the eos-sdk default branch to stable to ensure the correct branch is
chosen. This matches the configuration in eos-application-tools.
Probably this should be specified in the configuration on the server
side, too.

https://phabricator.endlessm.com/T22340