defaults: remove explicit specification of SDK 3
Now that we have ported the apps to SDK 4, we no longer want to include
SDK 3 by default.  In fact, even the base image will get the proper
SDK version as dependencies of the required apps, so let's stop
specifying any default platform runtime in addition to those that are
installed automatically as app dependencies.

https://phabricator.endlessm.com/T23220