eib: Add functions to delete loop devices for buildroot
If a loop device is active with a backing file inside a buildroot, then
it will prevent the buildroot from being deleted. The
delete_root_loops() function finds any loops the have backing files
within a specified root and deletes them. If there are any partitions on
the loop device then they'll be deleted, too. The assumption is that
the loop and any partitions are not mounted.

https://phabricator.endlessm.com/T22039