eib: Delete loops while unmounting filesystems
In order to delete the loop devices, we need to be sure they're not
mounted anymore. One way to ensure that is to delete the loop when the
path with its backing file is the last mount in the hierarchy. If all
the subsequent mounts have been unmounted, then we can be sure that any
mounts of the loop device have been unmounted.

https://phabricator.endlessm.com/T22039