config: Enable P2P app updates by default 🎉
This commit enables P2P (LAN/USB) app updates by default on all images,
only for Endless applications since Flathub isn't ready. While USB
updates are not quite ready to be a user-visible feature, the remaining
bugs are only an issue if you're trying to use USB updates and not if
you're doing Internet updates. So I think it's reasonable to go ahead
and enable collection IDs on newly produced images. Note that this
doesn't apply to old images that update; see T19363 for that case.

https://phabricator.endlessm.com/T22604