applist: group by runtime (by default)
The locale-specific vs. generic distinction was useful at the time, but
a more common question these days is "which apps would I have to drop to
get rid of org.gnome.Platform//3.24?"

This also addresses the problem where extensions (which are technically
runtimes) were listed in the "runtimes" category, rather than alongside
the apps/"real" runtimes that depend on them. One caveat: runtime
extensions are now not shown at all in the old, by-"nature" grouping.
This'll do for now.