config: Differentiate between upload and download URLs
The images.endlessm.com domain is for uploading images and querying
metadata about them. The images-dl.endlessm.com domain is for
downloading images.

Explicitly separate the two in the configuration, prefixing the existing
options with `upload_api__`, and changing the publish hooks to explicitly
use the appropriate upload or download URL depending on what they’re
trying to achieve.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23523