publish: Trigger OS update tests in OpenQA when configured
Support config options which would trigger update tests to and from the
image being built, in OpenQA. This allows us to test the OS update path
between different releases of the OS.

We use the image server API to query data about releases on each branch
we’re updating to/from, and download the config and manifest for every
such image in order to schedule the OpenQA jobs.

Prefer to get information about old builds from the manifest rather than
from the reconstructed branch config. The manifest is archived for each
image, so should accurately represent the configuration which went into
that image. The config is kept in the eos-image-builder repository, and
could also have been affected by local overrides on the machine which
did the image build — so it’s less reproducible.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23693