Include ostree-metadata refs in images
When collection IDs (and therefore P2P support) are enabled, ostree
stores repo metadata in contentless commits on the ref "ostree-metadata"
on the server side, and this ref is distributed by clients
participating in LAN/USB updates. However, since the image builder
currently sets collection IDs in the repo config at the last second
after doing all the other operations on the repo such as installing
apps, the ostree-metadata ref is never pulled from the server. The
effect of this is that if you install a recent Endless image on a
computer and never go online, you won't be able to copy your apps or OS
update to USB drives. Only after doing an online operation (such as an
app update) will you pull ostree-metadata from the server and be able to
copy your updates to USB drives. This is unfortunate since offline users
are exactly who the feature is aimed at.

So this commit changes the eib_image stage so that it explicitly pulls
the ostree-metadata ref into the repo after pulling the primary OS ref,
and changes eibflatpak.py so that the collection IDs are added to the
remote configs **before** doing appstream and remote config updates on
each remote, so those operations will pull in ostree-metadata refs.

One caveat is that while this happens unconditionally for the OS remote,
the ostree-metadata is only pulled for flatpak remotes that have a
collection ID configured. This effectively means every remote but
eos-runtimes at the moment, but since that one is not P2P enabled at the
moment anyway it doesn't seem worth doing a different implementation.

The reason we decided to configure the collection IDs at the end of the
image build was so that the server would only pull from the Internet not
LAN/USB sources, and so that it wouldn't exercise the experimental P2P
code. At this point the code is considered stable, but a side effect of
this commit is that the image build server will check LAN/USB sources
when updating the appstream/metadata, and presumably fail to find
anything in either. To prevent that behavior we probably need to
implement the fix for https://github.com/flatpak/flatpak/issues/1863

https://phabricator.endlessm.com/T23754