config: Enable P2P OS updates by default 🎉
This commit tells the image builder to add a collection ID to the OS
update remote, which is a necessary but insufficient requirement for
LAN/USB OS updates to happen. The other requirement is for the
eos-updater config file to specify LAN/USB as sources for updates. Until
now this has been done as an image builder hook that overrides the
config for some personalities, but moving forward it will be done in
eos-updater itself; see https://github.com/endlessm/eos-updater/pull/233

The USB OS update feature is slated for the 3.4.8 release, so this
commit makes that happen for newly flashed machines; enabling the
functionality in an update will be the purview of T19363.

Since the eosoem product changes this config key back to false, this
change does not apply there, as is the case for P2P app updates. The
eosoem product is being temporarily excluded from this feature as a way
to mitigate risk since the feedback loop for those images happens on
much longer timescales (see T22604).

https://phabricator.endlessm.com/T23413