publish: Expose OSTree stage information when plotting OpenQA updates
Some updates (for example, from any branch → master) require the system
under test to have eos-stage-ostree to be run on it before updating. In
order to do that, the test harness needs to know whether to change the
stage, what to change it to, and what password to use for it.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T23693