FlatpakRemote: Clear existing collection ID from cached remotes
Recent flatpak began automatically adding collection IDs when the remote
has the key ostree.deploy-collection-id in the summary. However, once
the collection ID is set in the remote configuration, flatpak will
refuse to change it again. This is a problem in the content repo that's
persistent across builds and where the remote URL can be changed by the
current build's configuration. When adding a remote, clear any
collection ID that might be present since it may not match the
configured URL.

https://phabricator.endlessm.com/T23863