FlatpakRemote: Delete ostree-metadata ref before updating remote
To support P2P updates, flatpak tries not to downgrade the
ostree-metadata ref if the remote commit is older than the current one.
However, the current ostree-metadata ref in the cached repo would be
from a previous build which may have been pointing to a different URL.
Delete the ref before updating the remote metadata to ensure it pulls
whatever is the latest commit on the remote.

https://phabricator.endlessm.com/T23863