Use images.endlessm-sf.com for publishing and querying
The images.endlessm.com DNS name points to a public address with strict
firewalling while the images.endlessm-sf.com DNS name points to an
internal address that's available anywhere on our internal network and
VPN. Since all the builders are required to be on the internal network,
use the internal DNS name. The public name will be dropped when there
are no more users.

https://phabricator.endlessm.com/T23750