publish: Add hook to block on S3 uploading
Normal publishing is done to the image server, which then pushes the
build to S3 on a timed jenkins job. In order to make the build block
until the assets have been pushed to S3, trigger the jenkins job and
then poll it until completion. This has been copied from
https://github.com/endlessm/eos-ostree-builder/blob/master/hooks/publish/60-sync-remote,
which has a similar 2 stage publishing.

The endless-ca-cert package is needed in the buildroot to trust the SSL
certificate at https://ci.endlessm-sf.com.

The hook is disabled by default since the authentication credentials are
private. Those will be provided by on the builder in
/etc/eos-image-builder/private.ini.

Eventually this will go away when publishing is done directly to S3 and
the image server does not have a local image store at all. That requires
some rework of the image server code, though.

https://phabricator.endlessm.com/T23658