eib_image: Change karg 'loglevel=2' to 'loglevel=0'
We previously added loglevel=2 to make the boot silent by default,
leaving only the critical messages displayed on screen. But it turns out
there are still many log messages that are printed on critical level or
bellow which don't bring us any useful information and prevent a clean
boot, for example MCE errors.

In addition grub now overrides the loglevel (appending a new kernel
parameter) on the boot following an unsuccessful boot, making it verbose
again.

https://phabricator.endlessm.com/T24262