Fix ARM builds if personality enables ISO images
It's not possible to build ISOs for our ARM platforms, so personalities which
enable iso = true override the iso = false in the ARM architecture try and
invoke the create_iso helper with a missing boot zip. Check for both [image]
iso and [image] boot_zip being true before invoking the ISO hook, and unset
boot_zip on ARM (already the case in practice, as the armhf path in
eib_image does not actually use this variable or create a boot.zip).

https://phabricator.endlessm.com/T24439