Add helpers to ignore chattr errors on overlayfs
When the image builder is running from a docker container, the
filesystem is overlayfs. At least on Ubuntu Xenial, overlayfs does not
support setting filesystem attributes on directories. That makes the
build fail when the OSTree deployment directory is attempted to be made
mutable.

OSTree already catches the ENOTTY and ENOTSUP errors when attempting to
make the deployment directory immutable, but our helpers to ensure the
directory mutability is restored just run bare chattr. Add a
mutable_path() utility function and mutable-path wrapper script that
also ignores these errno values. The ioctl and flags values have been
tested on various systems and seem to work correctly.

https://phabricator.endlessm.com/T25810