publish: Drop publishing to local host
Way back in the day we built and published images on the same server.
That hasn't been the case for years and the code to handle that is just
old cruft. Drop it and the short host name setting that was only there
to support it. If anyone wants to actually publish locally they can just
make sure ssh works.

https://phabricator.endlessm.com/T24316