publish: Upload build output directly to S3
Instead of telling the server via jenkins to upload the files to S3 and
then blocking in a fragile way until that's done, just send the files
there ourselves. This needs to be done before uploading the files to the
server so that it doesn't concurrently try to upload them to S3.

This obviously requires the image builder to have appropriate privileges
for writing objects in S3, but it already has credentials for copying
content from S3 buckets. The appropriate privileges just need to be
attached to the existing IAM user in AWS.

https://phabricator.endlessm.com/T24316