lib: Limit xz to 1GB of memory
xz preallocates a chunk of memory for each thread when compressing.
Since we specify -T0, xz will create a thread for each CPU. On our
packet ARM server there are 64 CPUs. In that case, xz fails immediately
with ENOMEM trying to allocate memory for 64 threads since that exceeds
the 32 bit addressable limit for armhf builds.

Pass the -M parameter to limit the total amount of memory. xz will limit
the amount of threads in use to fit the allocation within this boundary.
I've chosen 1GB as that still starts 12 threads. The -vv argument will
show exactly what xz has chosen to do. For example:

  + 23:25:27 eib.sh: xz -vv -M1G -T0 -4 -c /var/cache/eos-image-builder/tmp/eos-eos3.5-armhf-ec100.190301-230307.base.img
  xz: Filter chain: --lzma2=dict=4MiB,lc=3,lp=0,pb=2,mode=normal,nice=16,mf=bt4,depth=0
  xz: Using up to 64 threads.
  xz: 5319 MiB of memory is required. The limit is 1024 MiB.
  xz: Decompression will need 5 MiB of memory.
  xz: Adjusted the number of threads from 64 to 12 to not exceed the memory usage limit of 1024 MiB

https://phabricator.endlessm.com/T25810