config/defaults: drop lower bang-for-byte games
I am taking guidance from past discussion on
https://phabricator.endlessm.com/T18626 when I last had to cut down the
size of the image, as well as a general principle that we believe the
Endless Studios games, while large, are higher value than many of the
games we've previously shipped.

Many of the games being dropped are quite small. One key exception is
wz2100: it's large in and of itself (1.1GB). OpenSCAD is only 28MB;
these are the only two apps in the default set which require
org.kde.Platform//5.11. If you take into account sharing with //5.12
which is also in our images, //5.11 accounts for about 800 MB.

MegaGlest is 500 MB and has very mixed reviews.

https://phabricator.endlessm.com/T26235