hooks/split: delete Flatpak deploy/ refs on disk1
To quote <https://github.com/flatpak/flatpak/pull/2563>:

> When we deploy e.g. app/org.foo.bar/x86_64/stable, then we
> also create a deploy/app/org.foo.bar/x86_64/stable ref pointing to
> the latest deployed commit. We also remove it when an app is uninstalled.
>
> This means that a prune operation will not delete objects that are deployed
> (which would not save any space anyway). This is nice because this can
> happen for instance when you flatpak update --no-deploy.

This behaviour was introduced in Flatpak 1.1.3.

Unfortunately, the hook which uninstalls apps from disk1 does not use
the Flatpak API to uninstall apps. Instead, it removes everything from
/var/lib/flatpak besides the repo symlink, then removes all refs that
look Flatpak-y, then prunes the repo. As a result, all the deploy refs
are left, and the prune operation does not delete these objects.

One might reasonably argue that this script should be changed to
uninstall everything through the Flatpak API, then delete all the stray
files, and only then prune. But there may be reasons why it was done
this way, so make the simpler change to delete these deploy/ refs.

https://phabricator.endlessm.com/T26877