config: disable 90-email-success hook
The majority of traffic on the endless-dev-status@ mailing list is build
success emails from eos-image-builder and eos-ostree-builder. In the
past, I occasionally found these emails useful because they could be
easily searched; these days, we put enough information in the job
descriptions on
https://ci.endlessm-sf.com/view/Image/job/image-build-amd64/ that one
can search there instead. I believe these emails are not otherwise
particularly useful or widely-read.

The cost of these emails is that build *failure* emails are hard to spot
among the success emails. It is useful to get timely notification of
build failures so they can be investigated, but the signal-to-noise
ratio on the endless-dev-status@ mailing list makes following it
impractical.

Disable success emails. Failure emails will continue to be sent by
Jenkins.

https://phabricator.endlessm.com/T27165