config/default: Change top 8 languages in FBE language chooser
This changes the languages listed first in the FBE language chooser to
match Endless target markets. These languages are where we invest in
translations/apps most actively, sorted roughly by volume.

Originally implemented in gnome-initial-setup as part of
https://phabricator.endlessm.com/T18020 and since there's now a vendor
file key to customize this list, we prefer to implement it here to
reduce our divergence with upstream gnome-initial-setup.

https://phabricator.endlessm.com/T25162