Remove FBE branding welcome page
This customization no longer needs to be used for Hack, and the only
remaining user is EDUCA images which I have been told are not shipping.
So we are removing the feature from gnome-initial-setup for the time
being, to avoid having to rebase it every 6 months when no-one is using
it. This removes all the uses of the feature from image build
configurations.

https://phabricator.endlessm.com/T25162