eib: Set extra-languages key instead of languages
Instead of setting xa.languages in the repo config only while fetching
a remote's refs so that the desired subpaths of locale extensions are
fetched, we now set xa.extra-languages and leave it set so it can be
used while the image is in use. The difference is that flatpak uses
xa.languages as an exhaustive set of locales, whereas it combines
xa.extra-languages with the user/system defined locales. This means
users get locale extensions with both the language they chose and
the ones we anticipated they might need, and it means GNOME Software
can display apps suitable for both sets of languages.

https://phabricator.endlessm.com/T24214