default: add Clubhouse from eos-apps
A minimal Hack experience should be available
in all the Endless amd64 images.

The new Clubhouse contains all the components
like the Hacking Toolbox, the Unlock puzzle
and the Sound server and builds the Clippy
extensions as well.

We also include the System app and the
ProjectLibrary into the minimal version.

https://phabricator.endlessm.com/T28291