FlatpakManager: Fetch all languages for the cache repo
For the cache repo, set xa.languages to * so that all languages are
fetched. During install this will be filtered down by
xa.extra-languages, but subsequent builds may make use of languages that
this build didn't.

The setting is cleared after enumeration is complete in case subsequent
stable branch builds aren't expecting it.

https://phabricator.endlessm.com/T24214