eib: Add opt-in support to set extra-languages
Add an opt-in support to set extra-languages in images. Previously, the
xa.extra-languages key was always set with whatever was on the locales
variable, which only made sense for xa.languages, that was used to fetch
refs for flatpaks, then unsetting them right after the pull was finished,
when we don't really know what language the user will choose for the OS.

With the extra-languages, however, this led to the undesireable
behaviour where all languages were set to appear on GNOME-SOFTWARE for
some images, because this option wasn't unset after pulling refs. The
opt-in support fixes this behaviour by only using the multi-locale
feature on images that needs a special selection of languages/locales.

https://phabricator.endlessm.com/T24214