split: delete Flatpak mirror refs
This script attempts to remove all traces of Flatpak from the OS repo,
which ends up on disk1. Due to some confusion around how Flatpak names
refs when collection IDs are involved, these were previously left
kicking around on disk1, meaning the apps' and runtimes' data couldn't
be pruned.

We cannot use either the Flatpak or the OSTree APIs to do this:

- Flatpak.Installation.uninstall() only removes the classic ref
- OSTree.Repo.list_collection_refs() cannot be called from Python
  <https://github.com/ostreedev/ostree/issues/1958> due to a bug in
  pygobject <https://gitlab.gnome.org/GNOME/pygobject/issues/278>. This will
  be fixed by <https://gitlab.gnome.org/GNOME/pygobject/merge_requests/127>
  but we're not there yet.

Taking advantage of the fact that only Flatpaks have mirror refs, just
remove them all the old-fashioned way.

TODO: replace this workaround with one of the following, in increasing order
of preference:

(a) use the OSTree API to remove collection refs for all but the collection
    ID of the OS, once the pygobject bug is fixed
(b) use the Flatpak API to uninstall all Flatpaks, once Flatpak is fixed to
    more consistently handle mirror refs
(c) stop unifying the Flatpak and OS OSTree repos
    https://phabricator.endlessm.com/T19248

https://phabricator.endlessm.com/T28344