eibflatpak: optimise caching of pulls by avoiding static deltas
Flatpak remotes typically have a scratch delta for every commit, and in
this case OSTree decides between doing an object fetch or a delta fetch
based on whether the ref already exists (ie, the heuristic that an
object fetch, which is usually slower, might be faster if you have a
previous version of the ref). For us, we are often fetching
incrementally or want the local cache repository to be checked for
objects, so disable static deltas across the board.