hooks: fix putting Chrome into favorite-apps
In the past, we applied our Endless-specific favorite apps by patching
the schema. Since https://phabricator.endlessm.com/T25835 we moved to
using vendor overrides, as is right and proper. However, this hook to
replace Chromium with Chrome on free images was implemented by scraping
the schema, so in an attempt to replace Chromium with Chrome we actually
undid our vendor customization and ended up with Evolution, Rhythmbox,
Shotwell, Nautilus, App Center on the taskbar – and no browser!

Fix this by scraping the override file instead, and add a check that the
string being modified actually contains "chromium-browser", which would
have caught this oversight.

https://phabricator.endlessm.com/T28610