config: remove 3 large games & 4 large English apps
These 3 games were already excluded on English images; remove them from
others to get the size down a bit.

The (English) knowledge apps ended up ranked relatively low on the big
list o' apps; the Blender one because it is not useful without Blender
which we cannot ship on free images, Video Games because we can't ship
Unity; Raspberry Pi because we haven't tested they can actually be used
on Endless and most people do not own a Raspberry Pi, and the other two
because they are not directly related to our mission.

This is an interim measure while the full review is completed – the
images have ballooned even further due to Hack so we need to remove
something to at least not get any further away from 24 GB installed.

https://phabricator.endlessm.com/T28249
https://phabricator.endlessm.com/T26787
https://docs.google.com/spreadsheets/d/1VTuLuF2o3FOYAq69MkdNvzy5D9G6n4DeuUVv7wUtlE4