config: inherit default locales in en personality
We will ask OEMs to use the 'en' image as the "fallback" image when we
don't provide a relevant language-specific image, rather than 'sea' as
they do at present.

In support of this, we want the 'en' image to inherit the dozen or so
locales included in the 'base' image. (The alternative would be to build
an explicitly global image which differed only in these locales.)

https://phabricator.endlessm.com/T26787