Specify Hack apps in defaults
Now that these applications are supported on arm64, the odd one out is
armhf *not* including these apps. List them as installed by default on
all architectures, and take them back out again on armhf. This makes it
less likely that we will forget to ship a new Hack app on arm64 when we
add it to amd64.