Revise default & en apps to emphasise coding & creation
https://docs.google.com/spreadsheets/d/1VTuLuF2o3FOYAq69MkdNvzy5D9G6n4DeuUVv7wUtlE4/
contains a list of apps with neutral or English content, in decreasing
order of priority. They have been arranged to correspond to our new
priorities of providing (coding) education content and games aimed at a
somewhat older audience, and tools for creation, with some general
reference content; as opposed to prioritizing general reference material
and educational material for a younger audience.

While the language-neutral applications shipped on non-English images
will be different as a result of this change, we make no particular
effort to revise those: we don't really have coding education and
content creation tutorials for other languages, so we may as well
continue to ship what we do have.

https://phabricator.endlessm.com/T26787