Allow specifying a maximum size for the full image
We are defining a new personality, lite_en, destined for systems with a
single 32GB disk. This personality's size should be capped at 24GB,
allowing headroom for user data, new apps, and updates for apps and the
OS.

To enforce this, allow the configuration to specify a maximum size for
the image and fail if it would be exceeded. This is similar to the
existing [image] disk1_size, [image] disk2_size and [iso] max_size
fields.

https://phabricator.endlessm.com/T26787