Build ready-to-use VM images
Add a new helper, create-vm-image to generate VM images ready to be
imported into VirtualBox and VMWare. The format used is a ZIP archive
containing a OVF file, a VMDK virtual disk, and the manifest with SHA256
digest for both.

This feature can be enabled / disabled by the 'vm_image' option in the
'[image]' section of image configuration files.

Signed-off-by: Chris Chiu <chiu@endlessm.com>
Signed-off-by: João Paulo Rechi Vita <jprvita@endlessm.com>

https://phabricator.endlessm.com/T29325