config: Rework OSTree repo URLs to use internal server for pulling
Our CDN is not good, so when we do production ostree builds they often
fail with HTTP 500 errors from the edge servers. Instead, always pull
from our ostree server using the ostree.endlessm-sf.com domain. The
ostree and flatpak remote URLs for the deployment will continue to point
to the canonical public ostree.endlessm.com (except for master builds,
which continue to point to the ostree server using our public
origin.ostree.endlessm.com domain).

While here rework the config variables to hopefully be more
understandable and usable. With the additional variables it's now easier
to override in specific places without repeating other parts of the URL.
I'm hoping this will be more maintainable.

Note that the flatpak repos continue to use a single URL pointing to the
CDN. This will be changed in a subsequent commit.

https://phabricator.endlessm.com/T29611