eib_image: Make vm_image only work when arch=amd64
For all other architecutres, this setting will be a no-op. This
limitation is documented in config/default.ini and every other place
where 'vm_image' is changed.

This is needed because a product-arch-personality configuration is
currently not supported. Also, since we are only installing the
necessary tools (qemu-img) to the amd64 build environment, this
currently breaks image builds for other architectures.

https://phabricator.endlessm.com/T29325