FlatpakFullRef: Refactor attributes
Instead of decomposing the Flatpak.RemoteRef and metadata, pass them in
to FlatpakFullRef and provide properties to access the attributes
needed. This changes runtime to always provide the full ref, drops sdk,
which wasn't used anywhere, and changes extra_data to the boolean named
has_extra_data since that seems more accurate.

https://phabricator.endlessm.com/T27888