manifest: Include version information from appdata
All apps on Flathub are now required to include version information in
their appdata (with the exception of apps which have not been rebuilt
since this policy was introduced). Some "real" runtimes also have this
information, though not .Locale extensions etc.

This information is more useful than the ostree commit ID or date,
although unlike the commit ID it does not uniquely identify the included
version since you can publish a new Flatpak with the same version in its
appdata.

    "org.learningequality.Kolibri/x86_64/stable": {
      "name": "org.learningequality.Kolibri",
      "arch": "x86_64",
      "branch": "stable",
      "remote": "flathub",
      "commit": "f4604ad83d34eff6bd8de93514eb7d9ebae8331ec621e8a1f3a7dd9a758e538b",
      "size": 358990336,
      "ref": "app/org.learningequality.Kolibri/x86_64/stable",
      "date": "2020-05-29T20:49:39+0000",
      "version": "1.0",
      "runtime": "org.gnome.Platform/x86_64/3.36"
    },

If not present, the value will be null in the JSON:

    "org.learningequality.KALite/x86_64/eos3": {
      "name": "org.learningequality.KALite",
      "arch": "x86_64",
      "branch": "eos3",
      "remote": "eos-apps",
      "commit": "0aae977d5310ef1aab21cd45141313bd37b64764ee41e506da3e72eb86aa1292",
      "size": 734645760,
      "ref": "app/org.learningequality.KALite/x86_64/eos3",
      "date": "2020-04-22T00:24:41+0000",
      "version": null,
      "runtime": "com.endlessm.Platform/x86_64/eos3.2"
    },