lib: Support udevadm settle in chroot
Normally `udevadm` checks to see if it's running in a chroot and exits
with the message `Running in chroot, ignoring request.` if it is.
However, if the `udevd` control socket and queue file are available,
then this typically can work so long as there aren't big mismatches
between the host `udevd` and the `libudev` in the chroot.

To support that, mount `/run/udev` in the build root and refactor the
calls to `udevadm settle` so that they use a helper. The helper checks
for the existence of the socket and sets the `SYSTEMD_IGNORE_CHROOT`
environment variable necessary to skip the check for running in a
chroot.

https://phabricator.endlessm.com/T30938