config: Drop the eosdvd product and assets
This configuration was used to produce EOS DVDs for Brazil and
Indonesia, but we have no plans to ship these anymore. This commit drops
the configuration files and the assets used for its custom ISO boot
splash animation.

This commit also remove any references to eosdvd from the README file.

https://phabricator.endlessm.com/T31091