ImageConfigParser: Automatically namespace merged options
In order to make the merging work correctly, every config file needs to
namespace the merged options with some unique suffix. This is really an
implementation detail, though, and shouldn't be forced onto the user.

This adds a dedicated `read_config_file` method for `ImageConfigParser`
that automatically namespaces the merged options so that the config
files can just use `option_add` and `option_del` without worrying about
how the merging will take place.

https://phabricator.endlessm.com/T31160