config: Remove merged option suffixes
Use the automatic namespacing so that the configuration does not have to
manage potential conflicts. Note that all of the references to default
values have been changed to `_defaults` to match the actual file name.
If all the config files are going to be rewritten, we might as well fix
that historical nitpick.

https://phabricator.endlessm.com/T31160