ImageBuilder: Move many settings to defaults.ini
Only when the setting needs to be defined at runtime do we need to
manage configuration options in the builder. In all other cases, let
them be in the actual configuration layer. Primarily this cleans up the
ImageBuilder implementation to not have to keep track of what the
defaults should be.

https://phabricator.endlessm.com/T31160