ImageBuilder: Remove configdir from configuration
This path can't be changed in the configuration since the builder needs
to know it before loading the configuration. Nothing in the
configuration needs it, so let it be a simple attribute of the class.
While here, allow the configdir to be set as a keyword argument. This
won't be exposed on the CLI, but it allows the tests to not have to
monkey with the instance attributes.

https://phabricator.endlessm.com/T31160