Support local directory of configuration
In addition to relying just `/etc/eos-image-builder/config.ini` and
`config/local.ini` for configuration customization, allow an entire
directory to be supplied mimicing the `config` directory in tree.

If `-d|--localdir` is specified on the command line, then
`<localdir>/config` will be included in configuration loading after
`config` and `/etc` but before `config/local.ini`. The latter is likely
not needed anymore, but has been retained for very simple local
customization.

Later `<localdir>/data` and `<localdir>/hook` can be considered to
provide more local customization of the build process.

https://phabricator.endlessm.com/T31160