eib: Ensure keyring is in PGP stream format
In newer gnupg, importing a key is already made to a keybox rather than
a PGP stream, regardless of where `--keyring` says to store the data at.
The keybox format is internal to gnupg. When exporting the keys with
`--export`, gnupg will output the data in the standard PGP stream
format.

Currently the keyring is only used by debootstrap, which uses `gpgv` to
verify the `Release` file from the target repo. That seems to handle the
keybox fine. However, inspecting a keybox is difficult, which makes
testing the keyring generation hard.

https://phabricator.endlessm.com/T31160