ImageBuilder: Don't set environment variables in process
None of the python code needs the configuration environment variables
set. They're only needed for the `buildscript` shell script and
everything that runs from there.

Instead of setting the environment variables in the main process,
construct a new environment and run `buildscript` with it. This is
cleaner for the real code and for testing where there's no monkey
patching of the real environment needed.

https://phabricator.endlessm.com/T31160