Config: Don't pull from the origin server.
The BunnyCDN we used was flaky, so some builds used the origin server to
bypass it (mainly for dev images where the cache didn't have the
objects). Now with Fastly, we don't have the problem anymore. Keep the
behavior of pulling from the internal server instead of going outside to
the CDN, but don't use the origin server as there's little reason for
it.

https://phabricator.endlessm.com/T31186